% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiers.R
\name{add_date_cols}
\alias{add_date_cols}
\title{Check for v-dem date columns and add/modify them where applicable.}
\usage{
add_date_cols(df)
}
\arguments{
\item{df}{A v-dem data.frame in wide format.}
}
\value{
A data.frame similar to the input, but with date columns adjusted.
}
\description{
Checks for the given data.frame if the "year" column exists and if it is of type numeric.
If it does not exist it is created from the "historical_date" column and in any case set
to type numeric.
The function also checks for a "historical_date" column and creates it from the "year" column if
necessary and makes sure it is of type Date.
}
