% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{all_forks_done}
\alias{all_forks_done}
\title{Check all forked processes.}
\usage{
all_forks_done(pattern = "fork", forklist)
}
\description{
Waits for all forked processes to finish and checks their return values
}
