% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{all_identical}
\alias{all_identical}
\title{Check if arguments are \code{identical}}
\usage{
all_identical(...)
}
\arguments{
\item{...}{Objects to check}
}
\value{
A single boolean indicating whether all of the provided
arguments are identical to each other.
}
\description{
Given an initial function argument, \code{all_identical} loops
through the remaining arguments checking that they are
\code{\link{identical}}.
}
\examples{
all_identical(list("a", "b"), list("a", "b"))

all_identical("Not", "Equal")

}
