% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.R
\name{bfa_not_run}
\alias{bfa_not_run}
\title{Check what bfa jobs haven't been run yet and the status of their dependencies}
\usage{
bfa_not_run(db)
}
\description{
Check what bfa jobs haven't been run yet and the status of their dependencies
}
