% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coders.R
\name{calc_coder_done}
\alias{calc_coder_done}
\title{Calculate the number of coders done with their assigned surveys}
\usage{
calc_coder_done(version, survey_ids = NULL, cutoff = NULL)
}
\arguments{
\item{version}{The version of the dataset to calculate the completion rates for.}

\item{survey_ids}{A vector of survey IDs to include in the calculation.}

\item{cutoff}{A numeric value between 0 and 1 representing the completion cutoff.
#' @return A data.frame with the coder, survey, and country of coders who have completed their assigned surveys.}
}
\description{
Calculate the number of coders done with their assigned surveys
}
\details{
This function calculates which country-surveys coders have completed. This is calculated based on their assigned country-surveys (task table) and the number of questions answered (log table). It filters out coders who have not completed at least the specified cutoff percentage of their assigned surveys. The function returns a data.frame with coder_id, survey_id, and country_id, and survey_name.

# @examples
\dontrun{
coders_done <- calc_coder_done(version = 15, survey_ids = c(2057, 2058), cutoff = 0.95)
}
}
