% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coders.R
\name{calc_percent_done}
\alias{calc_percent_done}
\title{Calculate the percent of question coders have completed in their assigned surveys}
\usage{
calc_percent_done(version, survey_ids = NULL)
}
\arguments{
\item{version}{The version of the dataset to calculate the completion rates for.}

\item{survey_ids}{A vector of survey IDs to include in the calculation.
#' @return A DataFrame with the completion rates for each coder, survey, and country.}
}
\description{
Calculate the percent of question coders have completed in their assigned surveys
}
\details{
This function calculates the completion rates for coders based on their assigned surveys and the number of questions answered. 

# @examples
\dontrun{
coders_done <- calc_coder_done(version = 15, survey_ids = c(2057, 2058))
}
}
