% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{check_writeable}
\alias{check_writeable}
\alias{file.writeable}
\alias{dir.writeable}
\title{Check if writable}
\usage{
file.writeable(s)

dir.writeable(s)
}
\arguments{
\item{s}{CharacterVector path}
}
\description{
Return \code{TRUE} if either a directory or file is writeable.
}
\section{Functions}{
\itemize{
\item \code{file.writeable}: Check file

\item \code{dir.writeable}: Check directory
}}

\examples{
\dontrun{
file.writeable("example.txt")
dir.writeable("./")
}

}
