% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_cy.R
\name{collect_aggregation}
\alias{collect_aggregation}
\alias{collect_max}
\alias{collect_last}
\alias{collect_first}
\alias{collect_mean}
\title{Country-year aggregation helper functions}
\usage{
collect_max(v)

collect_last(v)

collect_first(v)

collect_mean(v)
}
\arguments{
\item{v}{Vector of any type.}
}
\description{
Collection of helper functions meant to used in conjuction with
\code{dplyr::summarise_all} when aggregating column-wise from the
country-date to country-year level.
}
\details{
When aggregating per year, the base R functions may return
    a different type than the original input vector. This will
    cause the \code{summarise} family of functions to error
    out. The collection of helper functions simply take care of
    missingness and \code{NA} as appropriate.
}
\section{Functions}{
\itemize{
\item \code{collect_max()}: Aggregate by max

\item \code{collect_last()}: Aggregate by last

\item \code{collect_first()}: Aggregate by first

\item \code{collect_mean()}: Aggregate by mean

}}
