% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_summarise.R
\name{colMedians}
\alias{colMedians}
\alias{colSDs}
\alias{rowMedians}
\alias{rowSDs}
\title{Row/Column-wise Median and Standard Deviation}
\usage{
colMedians(x)

colSDs(x)

rowMedians(x)

rowSDs(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A numeric vector
}
\description{
Calculate the median and standard deviation by column or row given
a numeric matrix. \code{NA}s will not be removed.
}
\examples{
m <- matrix(1:9, 3, 3)
m[1, 1] <- NA

colMedians(m)
rowSDs(m)

}
