% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{column_types}
\alias{column_types}
\title{Determine column types of a data.frame}
\usage{
column_types(df)
}
\value{
Named character vector of coulmn types.
}
\description{
Determine column types of data.frame
}
