% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.R
\name{components_running}
\alias{components_running}
\title{Check which components for an index are running}
\usage{
components_running(db, var, codebook, qtable)
}
\description{
Check which components for an index are running
}
