% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{country_trans}
\alias{country_trans}
\title{Display country transformations}
\usage{
country_trans(v, db_internal = pg_connect("vdem_data"), full = FALSE)
}
\arguments{
\item{countries}{Character or numeric vector.}

\item{db}{database connection.}
}
\description{
Given a vector of \code{country_ids}s, \code{country_text_ids}, and 
\code{country_names} returns all other available transformations.
}
\examples{
\dontrun{
country_options(c(4,"USA", "South Korea"))
}
}
