% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_component_plot}
\alias{create_component_plot}
\title{Creates components plots for D variables
Showing all components of the index (and rescaling where needed)}
\usage{
create_component_plot(components_df, component_vartype, index, qtable)
}
\description{
Creates components plots for D variables
Showing all components of the index (and rescaling where needed)
}
