% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_country_plot}
\alias{create_country_plot}
\title{Face validity check plots}
\usage{
create_country_plot(country_data, var, min_y, max_y)
}
\description{
Generates face validity check plots, comparing two different versions of the dataset
Can be used for C, D, A, and E variables as long as they are not text data.
}
