% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{duplicates}
\alias{duplicates}
\title{Show duplicates from a data.frame}
\usage{
duplicates(df, cols)
}
\arguments{
\item{df}{A data.frame.}

\item{cols}{A character vector of column names}
}
\description{
Show the duplicates of a data.frame based on spcified columns
The output data.frame shows the original value and the duplicates.
}
