% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_date.R
\name{get_date}
\alias{get_date}
\alias{get_text_id}
\title{Extract date or country from country-date}
\usage{
get_date(x, party = FALSE)

get_text_id(x, party = FALSE)
}
\arguments{
\item{x}{Character Vector}
}
\value{
Date vector
}
\description{
Extracts the individual components from a country-date string
}
\details{
Country-date strings are a string concatenation of a three
    letter country ID, \code{country_text_id}, and an ISO 8601
    date, \code{historical_date}. \code{get_date} is a vectorised
    function that returns the date part, converted using
    \code{as.Date}.

The \code{get_text_id} function is the vectorised
    counterpart to \code{get_date} that returns the
    \code{country_text_id} components
}
\examples{
get_date("SWE 1988-04-20")

get_text_id("DEU 1951-02-23")

}
\seealso{
Other country_date functions: 
\code{\link{get_party_id}()}
}
\concept{country_date functions}
