% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_date.R
\name{get_party_id}
\alias{get_party_id}
\title{Extract party_id from the string}
\usage{
get_party_id(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Integer vector of length 1 to 6 separated by space.
}
\description{
Extract party_id from the string
}
\examples{
get_cpid("RUS 2244 2016-09-18")
}
\seealso{
Other country_date functions: 
\code{\link{get_date}()}
}
\concept{country_date functions}
