% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{get_posterior}
\alias{get_posterior}
\title{Process posteriors for exporting}
\usage{
get_posterior(path)
}
\arguments{
\item{path}{File path to MM/BFA/Binary/HLI file with posteriors}
}
\description{
Process posteriors for exporting
}
\details{
Normally the posteriors are located in \code{super} (BFA, MM) or \code{posterior} (HLI) directories.
}
