% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_tags.R
\name{get_root}
\alias{get_root}
\title{Question tag root}
\usage{
get_root(x)
}
\arguments{
\item{x}{CharacterVector of question names. For example, the column
names from the final V-Dem DS.}
}
\description{
Given a vector of question names, \code{get_root} returns the
original, non-transformed corresponding tags. Note, ordinalized
versions of the HLIs are considered separate question tag, since
they have alternative entries in the codebook.
}
\section{Warning}{
 We will only transform V-Dem
variable tags, meaning that we'll attempt to match \code{v\\d}
or \code{e_v\\d}, while always excluding the direct democracy
variables (v2dd* or v2xdd*) since we never create separate
versions of those.
}

\examples{
get_root(c("v2clacfree", "v2clacfree_osp", "v2x_freexp_codehigh"))

}
