% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_tag}
\alias{get_tag}
\title{Get variable name from file path}
\usage{
get_tag(file)
}
\arguments{
\item{file}{CharacterVector}
}
\description{
\code{get_tag} returns the v2/v3 variable tag name given a file
path pointing to a single-variable rds. This is really only
intended for for OSP/ORD/summary output of C-vars.
}
\examples{
get_tag("prep/interpolated/v2clacfree.rds")
get_tag("ord/cy/v2elembaut_ord_cd.rds")

}
