% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.R
\name{hpc_status}
\alias{hpc_status}
\title{Quickly look at status of all jobs running on the hpc}
\usage{
hpc_status(hpc_alias = "tetralith", user = "x_linfo")
}
\arguments{
\item{hpc_alias}{The alias of the hpc to ssh into. Default is "tetralith"}

\item{user}{The user on the hpc to check their jobs. Default is "x_linfo"}
}
\description{
Quickly look at status of all jobs running on the hpc
}
