% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{label}
\alias{label}
\alias{set_spss_labels}
\alias{set_stata_labels}
\title{Variable labels}
\usage{
set_spss_labels(x, labels)

set_stata_labels(x, labels)
}
\arguments{
\item{x}{\code{DataFrame}}

\item{labels}{CharacterVector of labels for each column.}
}
\description{
Sets either Stata or SPSS variable labels for a data frame by
adjusting the attributes of the object.
}
\section{Functions}{
\itemize{
\item \code{set_spss_labels}: Set SPSS variable labels

\item \code{set_stata_labels}: Set Stata variable labels
}}

