% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data.R
\name{load_ext_source}
\alias{load_ext_source}
\title{Load and clean external data}
\usage{
load_ext_source(file_name = "wb_pop", con)
}
\arguments{
\item{file_name}{Full or relative path to the file}

\item{con}{PostgreSQL connection to the database with external data}
}
\value{
data.frame with V-Dem country_id, year, and target variable(s).
}
\description{
Read in the data, clean it and assign V-Dem country_id
}
\details{
Nothing more, nothing less than just reading the data, guessing the source,
and matching cleaning function and merging table for the source. It is not
vectorised. Before using make sure that your .pgpass file exists and has all
relevant fields (the function connects to V-Dem data database to extract information
about countries and country units).
}
\examples{
\dontrun{load_ext_source(list.files("~/Documents/edata/2021/upd/", full.names = TRUE)[1])}

}
