% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{locate_z.sample}
\alias{locate_z.sample}
\title{Locate and load a z.sample file}
\usage{
locate_z.sample(path = ".", pattern)
}
\arguments{
\item{path}{Directory location of z.sample files}

\item{pattern}{V2 variable tag pattern}
}
\description{
Finds the z.sample (posterior \code{Z} sample csv) matching the
given pattern.
}
\details{
Our z.sample files are tagged according to the number of
    iterations run in the MM model. We do this because we need to
    be able to match the various resulting MM output files to
    specific runs. The consequence though is that downstream code
    doesn't know the absolute file names when trying to load the
    \code{Z} posteriors for a given \code{C} variable.
}
\section{Warning}{
 \code{locate_z.sample} will error out if multiple
    files are matched, which is particularly a problem when there
    exists binary or recoded versions of the same variable. So be
    specific in the regex pattern provided.
}

