% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_msg}
\alias{log_msg}
\title{Formatted logging}
\usage{
log_msg(text, type = c("INFO", "WARN", "ERROR"))
}
\arguments{
\item{text}{Text, string to be printed.}

\item{type}{Type of message, either WARN or INFO.}
}
\description{
Outputs a formatted log message.
}
\examples{
log_msg("This is a warning", type = "WARN")

}
