% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_scheduler.R
\name{make_id}
\alias{make_id}
\title{Get the task_id for a new task}
\usage{
make_id(var_name, conn)
}
\arguments{
\item{var_name}{A character vector; it can be of any length.}

\item{conn}{A DBI connection to PostgreSQL database that has a schema called "pipe".}
}
\description{
Get the task_id for a new task
}
