% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_scheduler.R
\name{make_module}
\alias{make_module}
\title{Create modules for Make table in PostgreSQL}
\usage{
make_module(mod_name, var_name, make_df, conn)
}
\arguments{
\item{mod_name}{A character vector of module names that need to be added for a variable.}

\item{var_name}{A character vector of length 1 that specifies the variable for which the module is added.}

\item{make_df}{A data.frame with all tasks.}

\item{conn}{A PostgreSQL connection.}
}
\description{
Create modules for Make table in PostgreSQL
}
\examples{
# !!!Don't Run!!!
# Don't run this function in parallel, NEVER EVER.
# Otherwise, everything is going to be screwed up!
# Below, you see a typical use of the function.

# library(vutils)
# conn <- pg_connect("v902")
# vars <- c("v2eldonate", "v2peasbsoc", "v2psprbrch", "v2exdfcbhs")
# mode_name <- c("download_var", "cleaning", "clean_all", "cont_hist_merge")
# mk <- tbl(conn, dbplyr::in_schema("pipe", "make")) \%>\% collect(n = Inf)

# lapply(vars, function(var) {
   # make_module(mode_name, var, mk, conn) \%>\%
   # mutate(docopt = NA_character_,
   #     file = paste0(var, ".rds"),
   #     tbl = NA_character_,
   #     version = "v902",
   #     ts = NA) \%>\%
   # pg_append_table(., "pipe.make", conn)
# })

}
