% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.R
\name{mm_done}
\alias{mm_done}
\title{Check what mm jobs are done and additional convergence details}
\usage{
mm_done(db, codebook, qtable)
}
\description{
Check what mm jobs are done and additional convergence details
}
