% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.R
\name{mm_running}
\alias{mm_running}
\title{Check what mm jobs are running at what iteration on the hpc}
\usage{
mm_running(db, codebook, qtable)
}
\description{
Check what mm jobs are running at what iteration on the hpc
}
