% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{osp}
\alias{osp}
\title{Linearized ordinal scale transformation}
\usage{
osp(z, gamma_mu)
}
\arguments{
\item{z}{Extracted \code{Z} parameter as a matrix.}

\item{gamma_mu}{Extracted \code{gamma_mu} as a matrix.}
}
\value{
Numeric matrix with the same dimensions as our Z matrix.
}
\description{
Transforms latent trait estimates into the original interval
scale.
}
