% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{partial}
\alias{partial}
\title{Partial function application}
\usage{
partial(fn, ...)
}
\arguments{
\item{fn}{A function.}

\item{...}{Arguments to be applied.}
}
\description{
Partially applies a variable number of arguments to a given function.
}
\examples{
x <- c(1, 2, NA, 4)
f <- partial(mean, na.rm = TRUE)
f(x)

}
