% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{pg_connect}
\alias{pg_connect}
\title{Connect to the V-Dem DB}
\usage{
pg_connect(dbname = "XXXXXXX")
}
\value{
Returns an object of class \code{PostgreSQLConnection}.
}
\description{
Convenience function meant to connect directly to the V-Dem
Postgres Database. Calls \code{\link{pg_pass}} with the default
location of \code{pass_file}, which should include all connection
details. For more fine grain control, \code{\link{pg_pass}} should
be used directly, or alternatively if a \code{.pgpass} entry is
missing, \code{\link[DBI]{dbConnect}}, \code{\link[DBI]{dbDriver}},
and \code{\link[RPostgres]{Postgres}}.
}
