% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_matrix}
\alias{read_matrix}
\title{Read in a csv as a matrix}
\usage{
read_matrix(f, rownames = TRUE, drop.vignettes = TRUE, ...)
}
\arguments{
\item{f}{File to open}

\item{rownames}{Logical indicating whether the first column should
used as the rownames for the resulting output matrix.}

\item{drop.vignettes}{Logical, whether to drop vignette rows based
on the \code{\link{is_vignette}} function. This is only useful
within the context of \code{rownames} being \code{TRUE}.}

\item{...}{Additional arguments passed to \code{\link{read_file}}}
}
\description{
Convenience function for creating a matrix from reading in a csv.
}
\details{
\code{\link[data.table]{fread}} doesn't set the row names
    automatically which is annoying since we store our sample
    posterior distributions as CSV for archive purposes.
}
