% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{round_df}
\alias{round_df}
\title{Round the numeric columns of a data.frame}
\usage{
round_df(df, n)
}
\arguments{
\item{df}{A data.frame}

\item{n}{Decimal position for rounding}
}
\value{
A data.frame
}
\description{
Round the numeric columns of a data.frame
}
