% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_scheduler.R
\name{run_make}
\alias{run_make}
\title{Determine order of tasks to run and pass on to run_task function}
\usage{
run_make(tasks, n_cores = 1, db)
}
\description{
Determine order of tasks to run and pass on to run_task function
}
