% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_scheduler.R
\name{run_task}
\alias{run_task}
\title{Run individual task}
\usage{
run_task(
  id,
  module_name,
  question_name,
  cmd = "Rscript",
  script,
  docopt = "",
  export = NA_character_,
  open_blas_single_threaded = TRUE
)
}
\description{
Run individual task
}
