% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{s_union}
\alias{s_union}
\title{Set union for variable number of arguments}
\usage{
s_union(...)
}
\arguments{
\item{...}{Variable number of arguments of any type.}
}
\description{
\code{s_union} returns the unique elements from a set
\code{\link{union}} between all function arguments.
}
\section{Warning}{
 The resulting output object will not inherit the
user-defined attributes of the input objects.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Also, as a further note, input objects are expected to be of
the same type and are therefore exposed to R's normal system of
type coercion.
}\if{html}{\out{</div>}}
}

\examples{
s_union(letters[1:3], letters[2:4], letters[3:5])

}
