% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_date.R
\name{sort_text_id}
\alias{sort_text_id}
\title{Sort a vector of country-dates}
\usage{
sort_text_id(x, decreasing = F, na.last = T, party = FALSE)
}
\arguments{
\item{x}{Character Vector}

\item{decreasing}{Logical indicating whether to sort by decreasing
order.}

\item{na.last}{Logical indicating whether NAs should be appending
to the end or beginning of the sorted vector.}
}
\description{
Sorts a character vector of country-dates based on the
\code{country_text_id} prefixes.
}
\details{
Sorts first country-date strings followed by vignette
    identifiers. The fact that we append vignettes to the end is
    simply a convenience for when working with the MM code.
}
\examples{
v <- c("USA 1920-01-01", "MEX 1902-01-01", "MEX 1900-01-01",
       "pilot_1_v1", "pilot_2_v1")
sort_text_id(v)

}
\seealso{
Other country-date functions: 
\code{\link{is_country_date}()},
\code{\link{is_vignette}()}
}
\concept{country-date functions}
