% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{tbl_wide}
\alias{tbl_wide}
\title{Download V-Dem data from PostgreSQL in wide format for a variable}
\usage{
tbl_wide(question_df, db_schema, var, tbl_name = NULL)
}
\arguments{
\item{question_df}{a data.frame with the columns question_id and name.}

\item{var}{a character vector of length 1 specifying the name of the variable
to filter the data.}

\item{tbl_name}{optional argument specifying the name of the table from which
to download the data. The default value is NULL and in this case the name of
the table in PostgreSQL is equal to the name of the database.}

\item{df_schema}{a vector with two elements: name of the database and schema
from which you need to grab the data}
}
\description{
\code{tbl_wide} returns the data frame for specified variable from certain
PostgreSQL table in a certain database.
}
\examples{
\dontrun{
# Do not run
library(vutils)
library(dbplyr)
library(dplyr)

question_df <- read_file("~/proj/mm-prep/ROOT/refs/question_table.rds") \%>\%
    select(question_id, name)
db_schema <- c("v9", "dataset")
var <- "v2expathhs"
dfWide <- tbl_wide(question_df = question_df, db_schema = db_schema, var = var)
}

}
