% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{to_seq}
\alias{to_seq}
\title{Convert keys to an unordered balanced sequence}
\usage{
to_seq(keys, min = -1.5, max = 1.5)
}
\arguments{
\item{keys}{A vector of any type.}

\item{min}{Start value for the output sequence}

\item{max}{End value for the output sequence}
}
\description{
Given a vector of \code{keys}, \code{to_seq} will return a
balanced sequence of numeric values from \code{min} to
\code{max}. The values will correspond to the unique alphabetical
ordering of \code{keys} applied to the original ordering of
\code{keys}.
}
\details{
This function is used exclusively to set the priors for
    the vignettes. Currently, based on the order of the thresholds,
    we set the priors to be from -1.5 to 1.5.

    Vignette thresholds can be sorted by any identifier as long as
    we can determine the proper order using \code{sort}.
}
\examples{
to_seq(c("a", "d", "c", "a"))

}
\seealso{
Other prior functions: 
\code{\link{offset_diff}()},
\code{\link{weighted.rowMeans}()}
}
\concept{prior functions}
