% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_date.R
\name{to_year}
\alias{to_year}
\title{Date to year}
\usage{
to_year(date)
}
\arguments{
\item{date}{Date or character}
}
\value{
Integer
}
\description{
Return the four digit year from a Date object.
}
\examples{
v <- as.Date(c("1900-01-01", "1901-01-01"))
to_year(v)

}
