% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_pipe.R
\name{update_task_status}
\alias{update_task_status}
\title{Update task status of make table}
\usage{
update_task_status(status = "done", id = Sys.getenv("TASK_ID"), db)
}
\description{
Update task status of make table
}
