% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{weighted.rowMeans}
\alias{weighted.rowMeans}
\title{Weighted Row Means}
\usage{
weighted.rowMeans(x, y, na.rm = T)
}
\arguments{
\item{x}{Matrix to be summarised rowwise by the weighted
mean.}

\item{y}{Matrix of the same dimensions as x. Used as the
weights for calculating the weighted mean of \code{x}.}

\item{na.rm}{Logical. Whether missing values should be removed.}
}
\value{
NumericVector the same length as \code{nrow(x)}.
}
\description{
Calculates the weighted row means for a complex, numeric, integer,
or logical matrix.
}
\seealso{
\code{\link{rowMeans}}

Other prior functions: 
\code{\link{offset_diff}()},
\code{\link{to_seq}()}
}
\concept{prior functions}
