% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{write_file}
\alias{write_file}
\title{Write an object to file}
\usage{
write_file(
  o = NULL,
  f,
  fn = NULL,
  rotate = FALSE,
  msg = T,
  overwrite = T,
  dir_create = F,
  ...
)
}
\arguments{
\item{o}{An object of any type.}

\item{f}{File to write to.}

\item{fn}{Optional argument to explicity specify the function
\code{R} should use to write the given object to file.}

\item{\\dots}{Additional arguments passed to the found function.}
}
\description{
Convenience wrapper that calls the appropriate function to write an
object to file.
}
\details{
If no function is provided, \code{save_file} selects a
    function from a hardcoded list depending on the suffix of the
    provided file. If a function is not found, an error will be
    raised.
}
