% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_data.R
\name{load_ext_source}
\alias{load_ext_source}
\title{Load and clean external data}
\usage{
load_ext_source(file_name = "wb_pop", con)
}
\arguments{
\item{file_name}{Full or relative path to the file}

\item{con}{PostgreSQL connection to the database with external data}
}
\value{
data.frame with V-Dem country_id, year, and target variable(s).
}
\description{
Read in the data, clean it and assign V-Dem country_id
}
\examples{
\dontrun{load_ext_source(list.files("~/Documents/edata/2021/upd/", full.names = TRUE)[1])}

}
